+++ subcommand/envoy-entrypoint/command_unix_test.go
@@ -116,6 +116,8 @@ func TestRun(t *testing.T) {
 					meta := makeTaskMeta(
 						"some-app-container",
 						"consul-ecs-control-plane",
+						"consul-dataplane",
+						"some-aws-managed-container",
 					)
 
 					if atomic.LoadInt64(&ecsMetaRequestCount) < 2 {
@@ -152,7 +154,7 @@ func TestRun(t *testing.T) {
 					// Sanity check. We mock two requests with app container running, and the rest with the app container stopped.
 					require.GreaterOrEqual(r, atomic.LoadInt64(&ecsMetaRequestCount), int64(3))
 
-					t.Logf("Check the fake Envoy process exits")
+					r.Logf("Check the fake Envoy process exits")
 					proc, err := os.FindProcess(envoyPid)
 					require.NoError(r, err, "Failed to find fake Envoy process")
 					// A zero-signal checks the validity of the process id.
@@ -195,11 +197,17 @@ func TestRun(t *testing.T) {
 func makeTaskMeta(containerNames ...string) awsutil.ECSTaskMeta {
 	var containers []awsutil.ECSTaskMetaContainer
 	for _, name := range containerNames {
-		containers = append(containers, awsutil.ECSTaskMetaContainer{
+		container := awsutil.ECSTaskMetaContainer{
 			Name:          name,
 			DesiredStatus: ecs.DesiredStatusStopped,
 			KnownStatus:   ecs.DesiredStatusRunning,
-		})
+			Type:          "NORMAL",
+		}
+
+		if container.Name == "some-aws-managed-container" {
+			container.Type = "AWS MANAGED"
+		}
+		containers = append(containers, container)
 	}
 
 	return awsutil.ECSTaskMeta{